chore: move type resolver into its own file
